% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-banks.R
\docType{data}
\name{banks}
\alias{banks}
\title{Banking Failures During the Great Depression}
\format{
An object of class \code{rda}
}
\source{
\href{https://github.com/jrnold/masteringmetrics/blob/master/masteringmetrics/data/banks.rda}{Banks online source}
}
\usage{
data(banks)
}
\description{
During the early great depression, the St. Louis Fed in the 8th Federal Reserve District was anti-interventionist,
while the Atlanta Fed in the 6th Federal Reserve District was an advocate for helping
banks in need. Mississippi was divided in half by these districts, with it's southern
half in Atlanta's 6th District, and the upper half in St. Louis's 8th district.
This data describes the number of banks in business over time in the 6th and 8th Federal
Reserve Districts.
}
\references{
Richardson, Gary, and William Troost. “Monetary Intervention Mitigated Banking Panics during the Great Depression: Quasi‐Experimental Evidence from a Federal Reserve District Border, 1929–1933.” Journal of Political Economy 117, no. 6 (2009): 1031–73. https://doi.org/10.1086/649603.
}
\keyword{datasets}
