% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mlda.R
\docType{data}
\name{mlda}
\alias{mlda}
\title{Minimum Legal Drinking Age data}
\format{
An object of class \code{rda}
}
\source{
\href{https://github.com/jrnold/masteringmetrics/blob/master/masteringmetrics/data/mlda.rda}{MLDA online source}
}
\usage{
data(mlda)
}
\description{
Data is from Carpenter and Dobkin (2009), which studied the effect of
Minimum Legal Drinking Age laws on mortality rates as predicted by age.
}
\references{
Carpenter, Christopher and Carlos Dobkin.
“The Effects of Alcohol Consumption on Mortality:
Regression Discontinuity from the Minimum Drinking Age,”
American Economic Journal: Applied Econometrics, 2009, 1:1, 164-182.
}
\keyword{datasets}
