% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.qed_did.R
\name{plot.qed_did}
\alias{plot.qed_did}
\title{Generic plotting function for "qed_did" objects}
\usage{
\method{plot}{qed_did}(qed_did, ...)
}
\arguments{
\item{qed_did}{an object of class "qed_did"}
}
\value{
a ggplot object visualizing trend lines between the pre-treatment period and  the post-treatment period
amongst the treatment and control groups. The points signify mean values of the dependent variable pre-treatment and post-treatment.
}
\description{
This function serves as a generic plot method for objects of class "qed_did", a result of a call to \link[QEDinfR]{qed_did}.
}
\details{
Edits can be made to the plots using ggplot functions (such as changes to axis ticks, labels, etc).
}
\examples{

data(banks)
fit <- qed_did("num_banks_in_business", "year", 1930,  "bank_district", "6th district", df = banks)
plot(fit)

}
