% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.qed_rdd.R
\name{plot.qed_rdd}
\alias{plot.qed_rdd}
\title{Generic plotting function for "qed_rdd" objects}
\usage{
\method{plot}{qed_rdd}(rdd_ls, cutoff = 0, interaction = TRUE)
}
\arguments{
\item{cutoff}{numeric, the cutoff criterion; default is 0 as qed_rdd will by default transform
covariate \code{x} to \code{(x - cutoff_score)}}

\item{interaction}{boolean, whether or not to allow for slope to be different between treatment and control}

\item{df}{object of type \code{qed_rdd}}
}
\value{
a ggplot object
}
\description{
provides convenient plot method for qed_rdd objects
}
\details{
One can add and edit plot using the ggplot functions, such as using xlab or
ylab to change the x and y labels.

The function automatically plots a line at the discontinuity if you kept
the \code{qed_rdd} default of transforming the \code{x} variable to
center the discontinuity at zero; otherwise, you must specify the discontinuity yourself.

By default the plot allows for an interaction effect between \code{x} and the
treatment; one can however set this to false.
}
\examples{

data(mlda)
fit <- qed_rdd("mva", "agecell", 21, mlda)
plot(fit)
plot(fit, interaction = FALSE)

}
