% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qed_did.R
\name{qed_did}
\alias{qed_did}
\title{qed_did: Difference in Difference Designs}
\usage{
qed_did(y, x, event_time, treated_id, treated, df)
}
\arguments{
\item{y}{string, name of the outcome variable in dataframe}

\item{x}{string, name of the time variable}

\item{event_time}{numeric, when the event occurred}

\item{treated_id}{string, name of the identifying variable in df}

\item{treated}{vector of the same type as the identifier variable, indicates which identifiers were treated by the event}

\item{df}{a dataframe or tibble}
}
\value{
object of type \code{\link[QEDinfR]{qed_did}}
}
\description{
\code{\link[QEDinfR]{qed_did}} estimates basic difference in difference designs and provides a convenient plotting utility to visualize counter factual outcomes following the parallel trends assumption.
}
\details{
The function automatically generates the post-treatment and treated indicators for estimating the model, and one may generate plots with \code{\link[QEDinfR]{qed_did}} objects easily with its plot method.
The \code{\link[base]{summary.lm}} method provides the appropriate summary statistics.
}
\examples{
data(banks)
fit <- qed_did("num_banks_in_business", "year", 1930,  "bank_district", "6th district", df = banks)
summary(fit)

}
\seealso{
\code{\link[stats]{lm}}
}
