% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qed_match.R
\name{qed_match}
\alias{qed_match}
\title{qed_match: Greedy Matching (with Propensity Scores or Mahalanobis Distances)}
\usage{
qed_match(
  formula,
  outcome,
  data,
  distance,
  caliper = NULL,
  replace = FALSE,
  ...
)
}
\arguments{
\item{formula}{formula, in the format: treatment_variable ~ c(covariates)}

\item{outcome}{string, name of outcome variable of interest}

\item{data}{dataframe or tibble, data to be used for matching}

\item{distance}{string, either "mahalanobis" or "propensityscore" to indicate the distance measure (i.e. criterion mataching is performed on)}

\item{caliper}{numeric, indicates the maximum possible distance a match can be made on (by default, NULL)}

\item{replace}{logical, indicates whether or not matching must be performed with replacement or without replacement (by default, FALSE)}
}
\value{
an object of class \code{qed_match} that includes outputs from the Match function, along with:
\describe{
\item{\strong{Matched}}{a dataframe containing matched observations}
\item{\strong{Data}}{a dataframe containing observations prior to matching}
}
}
\description{
Performs Greedy Matching (Nearest Neighbor Matching) in order to balance covariates in observational data
for causal inference.
}
\details{
This function serves as a wrapper for the \code{\link[Matching]{Match}} function in the \strong{Matching} package, and restricts distance measures to Mahalanobis distances
and Propensity Scores (calculated through a logistic regression). It aims to balance the treatment and control groups on all included covariates in order to perform causal analysis.
}
\examples{
\dontrun{
data(lalonde)
# For matching with propensity scores:
ps_match <- qed_match(treat ~ age + educ + black, "re78", lalonde, "propensityscore", caliper = NULL, replace = FALSE)
# For matching with Mahalanobis distances:
md_match <- qed_match(treat ~ age + educ + black, "re78", lalonde, "mahalanobis", caliper = NULL, replace = FALSE)
}
}
\seealso{
\code{\link[Matching]{Match}},
\code{\link[QEDinfR]{summary.qed_match}},
\code{\link[QEDinfR]{test.qed_match}}
}
