% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qed_rdd.R
\name{qed_rdd}
\alias{qed_rdd}
\title{qed_rdd: Regression Discontinuity Design}
\usage{
qed_rdd(y, x, cutoff_criterion, df, transform = TRUE, interaction = TRUE)
}
\arguments{
\item{y}{string, name of depedent variable/outcome}

\item{x}{string, name of independent variable}

\item{cutoff_criterion}{number, indicates where the discontinuity in x occurs}

\item{df}{A dataframe or tibble}

\item{transform}{boolean, whether or not to substract cutoff criterion from \code{x}}

\item{interaction}{boolean, whether or not to include an interaction between treatment and the covariate \code{x}}
}
\value{
an object of type \code{qed_rdd}
}
\description{
Estimates a simple Regression Discontinuity Design model with convenient methods
for summary and plotting
}
\details{
text
}
\examples{
\dontrun{
data(mlda)
fit <- qed_rdd("all", "agecell", 21, mlda)
summary(fit)
}
}
\seealso{
\code{\link[stats]{lm}} \code{\link[QEDinfR]{plot.qed_rdd}}
}
