% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qed_rdd.R
\name{qed_rdd}
\alias{qed_rdd}
\title{qed_rdd: Regression Discontinuity Design}
\usage{
qed_rdd(y, x, cutoff_criterion, df, transform = TRUE, interaction = TRUE)
}
\arguments{
\item{y}{string, name of dependent variable/outcome}

\item{x}{string, name of independent variable}

\item{cutoff_criterion}{numeric, indicates where the discontinuity in x occurs}

\item{df}{dataframe or tibble}

\item{transform}{logical, whether or not to substract cutoff criterion from \code{x}}

\item{interaction}{logical, whether or not to include an interaction between treatment and the covariate \code{x}}
}
\value{
an object of type \code{qed_rdd}
}
\description{
Estimates a simple Regression Discontinuity Design model with convenient methods
for summary and plotting
}
\details{
code{QED_rdd} estimates basic regression discontinuity designs and offers a convenient plotting method to display the discontinuous relationship.
The function automatically transforms the discontinuous covariate to center the discontinuity at zero.
By default it allows for an interaction effect.
}
\examples{
data(mlda)
fit <- qed_rdd("all", "agecell", 21, mlda)
summary(fit)

}
\seealso{
\code{\link[stats]{lm}} \code{\link[QEDinfR]{plot.qed_rdd}}
}
